reorganize dataset endpoint handlers
